Bug fixes/general maintenance

+++ AssignmentOne/src/app/nav/nav.component.ts
@@ -15,9 +15,13 @@ export class NavComponent implements OnInit {
 
   logout(){
     event.preventDefault();
-    localStorage.clear();
-    console.log(localStorage);
-    alert('You are now logged out!');
-    this.router.navigateByUrl('');
+    if (!localStorage.username || !localStorage.email){
+      alert('You are already logged out!');
+    }else{
+      localStorage.clear();
+      console.log(localStorage);
+      alert('You are now logged out!');
+      this.router.navigateByUrl('');
+    }
   }
 }
